refactor: extract shared helpers, components, and break up Admin page
Server:
- Extract DB helpers (getTotalPot, getGameById, getGameByPosition,
  calculatePayoutAmount, columnExists) to reduce query duplication
- Update standings, bracket, socket, and admin routes to use shared helpers
- Split closeAuction() into focused helper functions
- Deduplicate CORS config in index.js

Client:
- Centralize REGION_COLORS, ROUND_NAMES, REGIONS, and api() wrapper in utils.js
- Update Auction, Bracket, MyTeams, and Standings to use shared imports
- Extract ParticipantAvatar component (used in 6+ places)
- Add compact mode to CountdownTimer (replaces inline TimeLeft in Admin)
- Break Admin.jsx (1082 lines) into 7 tab files under pages/admin/
